chore: Refactor contents of `storage-browser/` path into `internals/`
