+++ packages/storage/__tests__/internals/apis/_getDataAccess.test.ts
@@ -3,9 +3,9 @@
 
 import { CredentialsProviderOptions } from '@aws-amplify/core/internals/aws-client-utils';
 
-import { getDataAccess } from '../../../src/storageBrowser/apis/getDataAccess';
+import { _getDataAccess } from '../../../src/internals/apis/_getDataAccess';
 import { getDataAccess as getDataAccessClient } from '../../../src/providers/s3/utils/client/s3control';
-import { GetDataAccessInput } from '../../../src/storageBrowser/apis/types';
+import { GetDataAccessInput } from '../../../src/internals/types/inputs';
 
 jest.mock('../../../src/providers/s3/utils/client/s3control');
 
@@ -57,7 +57,7 @@ describe('getDataAccess', () => {
 
 	it('should invoke the getDataAccess client correctly', async () => {
 		expect.assertions(6);
-		const result = await getDataAccess(sharedGetDataAccessParams);
+		const result = await _getDataAccess(sharedGetDataAccessParams);
 
 		expect(getDataAccessClientMock).toHaveBeenCalledWith(
 			expect.objectContaining({
@@ -96,7 +96,7 @@ describe('getDataAccess', () => {
 			$metadata: {},
 		});
 
-		expect(getDataAccess(sharedGetDataAccessParams)).rejects.toThrow(
+		expect(_getDataAccess(sharedGetDataAccessParams)).rejects.toThrow(
 			'Service did not return valid temporary credentials.',
 		);
 	});
@@ -110,7 +110,7 @@ describe('getDataAccess', () => {
 			$metadata: {},
 		});
 
-		const result = await getDataAccess({
+		const result = await _getDataAccess({
 			...sharedGetDataAccessParams,
 			scope: MOCK_OBJECT_SCOPE,
 		});
