+++ packages/storage/src/internals/apis/_getDataAccess.ts
@@ -10,12 +10,15 @@ import { CredentialsProviderOptions } from '@aws-amplify/core/internals/aws-clie
 import { getStorageUserAgentValue } from '../../providers/s3/utils/userAgent';
 import { getDataAccess as getDataAccessClient } from '../../providers/s3/utils/client/s3control';
 import { StorageError } from '../../errors/StorageError';
+import { GetDataAccessInput } from '../types/inputs';
+import { GetDataAccessOutput } from '../types/outputs';
 import { logger } from '../../utils';
+import { DEFAULT_CRED_TTL } from '../utils/constants';
 
-import { GetDataAccessInput, GetDataAccessOutput } from './types';
-import { DEFAULT_CRED_TTL } from './constants';
-
-export const getDataAccess = async (
+/**
+ * @internal
+ */
+export const _getDataAccess = async (
 	input: GetDataAccessInput,
 ): Promise<GetDataAccessOutput> => {
 	const targetType = input.scope.endsWith('*') ? undefined : 'Object';
