+++ packages/storage/src/internals/locationCredentialsStore/constants.ts
