+++ packages/storage/src/internals/managedAuthConfigAdapter/createListLocationsHandler.ts
@@ -1,8 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { CredentialsProvider, ListLocations } from '../types';
-import { listCallerAccessGrants } from '../apis/listCallerAccessGrants';
+import { CredentialsProvider, ListLocations } from '../types/credentials';
+import { _listCallerAccessGrants } from '../apis/_listCallerAccessGrants';
 
 interface CreateListLocationsHandlerInput {
 	accountId: string;
@@ -14,7 +14,7 @@ export const createListLocationsHandler = (
 	handlerInput: CreateListLocationsHandlerInput,
 ): ListLocations => {
 	return async function listLocations(input = {}) {
-		const result = await listCallerAccessGrants({ ...input, ...handlerInput });
+		const result = await _listCallerAccessGrants({ ...input, ...handlerInput });
 
 		return result;
 	};
