+++ packages/storage/src/internals/managedAuthConfigAdapter/index.ts
