--- packages/storage/src/internals/types/index.ts
@@ -1,2 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
