+++ packages/storage/src/internals/types/outputs.ts
@@ -1,0 +1,14 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { ListLocationsOutput, LocationCredentials } from './credentials';
+
+/**
+ * @internal
+ */
+export type ListCallerAccessGrantsOutput = ListLocationsOutput;
+
+/**
+ * @internal
+ */
+export type GetDataAccessOutput = LocationCredentials;
