--- packages/storage/src/storageBrowser/index.ts
@@ -1,25 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export { LocationCredentialsProvider } from '../providers/s3/types/options';
-export { StorageSubpathStrategy } from '../types/options';
-
-export { createLocationCredentialsStore } from './locationCredentialsStore';
-export {
-	AuthConfigAdapter,
-	createManagedAuthConfigAdapter,
-	CreateManagedAuthConfigAdapterInput,
-} from './managedAuthConfigAdapter';
-export {
-	GetLocationCredentials,
-	ListLocations,
-	LocationCredentialsStore,
-	CreateLocationCredentialsStoreInput,
-	LocationCredentials,
-	ListLocationsInput,
-	ListLocationsOutput,
-	GetLocationCredentialsInput,
-	GetLocationCredentialsOutput,
-	Permission,
-} from './types';
-export { AWSTemporaryCredentials } from '../providers/s3/types/options';
